+++ tests/problems/test_circuit_problem.py
@@ -13,9 +13,9 @@
 
 import pytest
 
-from nucs.constants import STATS_IDX_ALG_BC_NB, STATS_IDX_SOLUTION_NB, STATS_IDX_SOLVER_CHOICE_NB
 from nucs.problems.circuit_problem import CircuitProblem
 from nucs.solvers.backtrack_solver import BacktrackSolver
+from nucs.statistics import STATS_IDX_ALG_BC_NB, STATS_IDX_SOLUTION_NB, STATS_IDX_SOLVER_CHOICE_NB
 
 
 class TestCircuitProblem: